backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
pydantic==2.10.0
httpx==0.27.2
python-multipart==0.0.12
requests==2.32.3
PyMuPDF==1.24.0
pdfplumber==0.11.0
Pillow==10.4.0

